# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leave-types:
    get:
      tags: [leave]
      summary: List leave types
      operationId: listLeaveTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaveType" }
    post:
      tags: [leave]
      summary: Create leave type
      operationId: createLeaveType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveTypeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveType" }
  /leave-types/{id}:
    get:
      tags: [leave]
      summary: Get leave type
      operationId: getLeaveType
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveType" }
    patch:
      tags: [leave]
      summary: Update leave type
      operationId: updateLeaveType
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveTypeInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveType" }
    delete:
      tags: [leave]
      summary: Delete leave type
      operationId: deleteLeaveType
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: No Content

  # ── Leave Balances ────────────────────────────────────────────────────────
...
  /pay-components:
    get:
      tags: [payroll]
      summary: List pay components
      operationId: listPayComponents
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: applicableTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PayComponent" }
    post:
      tags: [payroll]
      summary: Create pay component
      operationId: createPayComponent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayComponentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayComponent" }
  /pay-components/{id}:
    get:
      tags: [payroll]
      summary: Get pay component
      operationId: getPayComponent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayComponent" }
    patch:
      tags: [payroll]
      summary: Update pay component
      operationId: updatePayComponent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayComponentInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayComponent" }
    delete:
      tags: [payroll]
      summary: Delete pay component
      operationId: deletePayComponent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: No Content

  # ── Payroll Periods ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    LeaveType:
      type: object
      properties:
        id: { type: number }
        codeEn: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        descriptionEn: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        category: { type: string }
        defaultDaysPerYear: { type: number }
        accrualFrequency: { type: string }
        accrualAmount: { type: string }
        maxCarryoverDays: { type: number }
        requiresApproval: { type: boolean }
        requiresAttachment: { type: boolean }
        minAdvanceNoticeDays: { type: number }
        maxConsecutiveDays: { type: ["number", "null"] }
        applicableToGender: { type: ["string", "null"] }
        isActive: { type: boolean }
        color: { type: string }
        createdAt: { type: string }
      required: [id, codeEn, nameEn, nameAr, category, defaultDaysPerYear, accrualFrequency, accrualAmount, maxCarryoverDays, requiresApproval, requiresAttachment, minAdvanceNoticeDays, isActive, color, createdAt]
    LeaveTypeInput:
      type: object
      properties:
        codeEn: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        descriptionEn: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        category: { type: string }
        defaultDaysPerYear: { type: number }
        accrualFrequency: { type: string }
        accrualAmount: { type: string }
        maxCarryoverDays: { type: number }
        requiresApproval: { type: boolean }
        requiresAttachment: { type: boolean }
        minAdvanceNoticeDays: { type: number }
        maxConsecutiveDays: { type: ["number", "null"] }
        applicableToGender: { type: ["string", "null"] }
        isActive: { type: boolean }
        color: { type: string }
      required: [codeEn, nameEn, nameAr]

    # ── Leave Balances ────────────────────────────────────────
...
    PayComponent:
      type: object
      properties:
        id: { type: number }
        codeEn: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        type: { type: string }
        calculationMethod: { type: string }
        value: { type: string }
        percentageBase: { type: ["string", "null"] }
        isTaxable: { type: boolean }
        isMandatory: { type: boolean }
        applicableTo: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, codeEn, nameEn, nameAr, type, calculationMethod, value, isTaxable, isMandatory, applicableTo, isActive, sortOrder, createdAt]
    PayComponentInput:
      type: object
      properties:
        codeEn: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        type: { type: string }
        calculationMethod: { type: string }
        value: { type: number }
        percentageBase: { type: ["string", "null"] }
        isTaxable: { type: boolean }
        isMandatory: { type: boolean }
        applicableTo: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: number }
        notes: { type: ["string", "null"] }
      required: [codeEn, nameEn, nameAr, type]

    # ── Payroll Periods ───────────────────────────────────────
...