# lib/api-zod/src/generated/types/leaveType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveType {
  id: number;
  codeEn: string;
  nameEn: string;
  nameAr: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  category: string;
  defaultDaysPerYear: number;
  accrualFrequency: string;
  accrualAmount: string;
  maxCarryoverDays: number;
  requiresApproval: boolean;
  requiresAttachment: boolean;
  minAdvanceNoticeDays: number;
  /** @nullable */
  maxConsecutiveDays?: number | null;
  /** @nullable */
  applicableToGender?: string | null;
  isActive: boolean;
  color: string;
  createdAt: string;
}